requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colormath==3.0.0
colour-science==0.3.16
cryptography==46.0.3
defusedxml==0.7.1
distro==1.9.0
docopt-ng==0.9.0
docx2txt==0.8
et_xmlfile==2.0.0
ExifRead==2.3.2
Flask==3.1.2
flask-cors==6.0.1
froeling-connect==0.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.46.0b1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
narwhals==2.12.0
natsort==8.4.0
networkx==3.5
numba==0.63.0b1
numpy==2.3.5
oauthlib==3.3.1
openai==2.8.1
openpyxl==3.1.5
organize-tool==3.3.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251107
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
PuLP==3.3.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.2.5
pypdf==6.4.0
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
reportlab==4.4.5
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.29.0
rsa==4.9.1
ruff==0.14.10
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
shap==0.50.0
simplematch==1.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
streamlit==1.52.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.4
xlsxwriter==3.2.9
yarl==1.22.0
youtube-transcript-api==1.2.3
